[REFACT] MMSE enum struct 파일 분리
